requirements.txt
requests>=2.31
Pillow>=10.0
imagehash>=4.3
PyYAML>=6.0
hypothesis>=6.0
pytest>=7.0
torch>=2.1
torchvision>=0.16
ultralytics>=8.0
scikit-learn>=1.3
scipy>=1.11
sagemaker>=2.200
